requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.23
langsmith==0.0.87
langchain-text-splitters==0.0.1
faiss-cpu==1.13.2
sentence-transformers==2.3.1
openai==1.10.0
httpx<0.28
python-dotenv==1.0.0
pydantic>=2.10.0
flask==3.0.0
